backend/requirements.txt
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.15
docstring_parser==0.16
executing==2.0.1
Flask==3.0.2
Flask-Caching==2.3.0
Flask-Cors==5.0.0
fpdf==1.7.2
google-ai-generativelanguage==0.6.6
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.72.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
Markdown==3.7
markdown-it-py==3.0.0
markdown_pdf==1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.6.0
node==1.2.1
nodejs==0.1.1
npm==0.1.1
numpy==1.26.4
odict==1.9.0
opentelemetry-api==1.28.1
optional-django==0.1.0
packaging==23.2
panadas==0.2
pandas==2.2.3
pandoc==2.4
parso==0.8.3
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.0
plumber==1.7
plumbum==1.9.0
ply==3.11
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pypandoc==1.14
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
setuptools==69.1.1
shapely==2.0.4
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.11.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wkhtmltopdf==0.2
wrapt==1.16.0
zipp==3.21.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.2
zope.lifecycleevent==5.0
zope.proxy==5.2
